# apps/api/public/openapi.template.yaml
openapi: 3.0.3
info:
  title: Semantic Score Calculator
  x-summary: API to compute the semantic score of an OpenAPI specification document.
  description: |-
    ## 🇬🇧 Compute the Semantic Score of an API

    This API computes the semantic score of an OpenAPI specification document
    looking at the mapping between RDF properties and JSON Schema properties
    defined via the `x-jsonld-context` OpenAPI 3.0 keyword.

    The Semantic Score is a value between 0 and 1, providing
    a synthetic indicator of the compliance of a data model
    to a given set of semantic assets published via schema.gov.it

    It can be used by developers, API consumers or API providers
    to gain insights into the data models used in their APIs
    and improve their semantic quality.

    For further information on mapping JSON Schema properties
    to RDF properties, see [REST API Linked Data Keywords](https://datatracker.ietf.org/doc/draft-polli-restapi-ld-keywords/).

    This API is associated with the Schema Editor provided by schema.gov.it
    at <https://teamdigitale.github.io/dati-semantic-schema-editor/latest/>

    ## 🇮🇹 Calcola il Semantic Score di una API

    Questa API calcola il punteggio semantico di un documento OpenAPI 3.0
    guardando il mapping tra proprietà RDF e proprietà JSON Schema
    definite tramite la keyword `x-jsonld-context` OpenAPI 3.0.

    Il Semantic Score è un valore decimale incluso tra 0 e 1,
    calcolato come rapporto tra il numero di proprietà semantiche risolte
    sui contenuti del Catalogo e quelle totali presenti nel file di definizione degli schemi dati.

    Può essere utilizzato da sviluppatori, consumatori e provider di API
    per ottenere informazioni sulle modalità di utilizzo dei modelli di dati
    e migliorare la qualità semantica delle API.

    Per ulteriori informazioni sul mapping tra proprietà JSON Schema e proprietà RDF,
    vedere [REST API Linked Data Keywords](https://datatracker.ietf.org/doc/draft-polli-restapi-ld-keywords/).

    Questa API è associata all'Editor di Schemi Dati fornito da schema.gov.it
    all'indirizzo <https://teamdigitale.github.io/dati-semantic-schema-editor/>

  version: 0.0.1
  contact:
    name: Team Digitale
    url: https://teamdigitale.governo.it
    email: teamdigitale@governo.it
  license:
    name: Apache-2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: https://schema.gov.it/note-legali
tags:
  - name: Health
    description: Know the health status of the service.
  - name: SemanticScore
    description: Compute the semantic score.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
    x-sandbox: true
  - url: https://example.com/api/v1
    description: Public server
    x-sandbox: false
security:
  # Anyone can use this API.
  - {}
paths:
  /status:
    get:
      summary: Retrieve the status of the service.
      tags:
        - Health
      operationId: HealthController_getHealth
      description: |-
        Returns whether the service is functioning properly or not.

        Consumers should use this endpoint to test the service,
        instead of issuing specific requests to the other endpoints.

        If the service is unavailable, the response may provide information
        about the unavailability period.
      parameters: []
      responses:
        "200":
          description: Service is healthy
          headers:
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                status: 200
                title: OK
        "429":
          description: Too many requests sent to the server.
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: |-
            The service is not available:
            further availability information may be
            provided via headers.
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          description: |-
            A client or server error happened while processing
            the service status.
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /semantic-score:
    post:
      summary: Calculate semantic score for an OAS 3.0 document.
      tags:
        - SemanticScore
      description: |-
        Process an OpenAPI 3.0 specification document containing
        a schema and computes its Semantic Score
        using the associated REST API Linked Data Keywords.

        The response contains the Semantic Score, that is a
        numeric value between 0 and 1.
      operationId: SemanticScoreController_updateSchemaSemanticScore_v1
      parameters: []
      security:
        # Anyone can use this endpoint.
        - {}
      requestBody:
        required: true
        description: |-
          The OpenAPI specification document containing
          a schema annotated with the REST API Linked Data Keywords.
        content:
          multipart/form-data:
            # The file should be in the following formats.
            encoding:
              file:
                contentType: application/openapi+yaml, application/openapi+json
            schema:
              $ref: "#/components/schemas/CalculateSemanticScoreRequest"
      responses:
        "200":
          description: |-
            The provided OpenAPI specification document has been successfully processed.
            The response will output information about the global semantic score
            and the models processed.
          headers:
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SemanticScoreResponse"
        "406":
          description: |-
            The provided file is not a valid OpenAPI 3.0 specification document.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "413":
          description: |-
            The provided file is too large.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "415":
          description: |-
            The provided file has an unsupported media type:
            we only accept application/json and application/yaml.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: |-
            Too many requests sent to the server.
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "default":
          description: |-
            Client or server error during semantic score calculation.
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
components:
  headers:
    Cache-Control:
      description: |-
        Cache usage SHOULD be extensively detailed in the `description` property
        to avoid data leaks or the use of stale data.
        This rule should ensure in some way that the API provider
        documents extensively the cache usage to avoid data leaks
        or use of stale data.
        For now this ruleset tests:
        * the presence of the following keywords
          in the `description`: `max-age`, `private`, `no-store`, `no-cache`.
        * that one and only one between Expires and Cache-Control is used.
        `Cache-Control` and `Expires` should not be used in conjunction,
        because `Cache-Control` overrides `Expires` when `max-age` is set.
        Instead, if neither `Cache-Control` nor `Expires` are set, clients MAY use heuristic cache
        as described in RFC7234.
      schema:
        example: no-store
        externalDocs:
          url: https://datatracker.ietf.org/doc/html/rfc7234
        pattern: >-
          $[a-zA-Z0-9=;" -]{0,255}
        type: string
    Retry-After:
      description: |-
        Retry contacting the endpoint *at least* after seconds.
        See https://tools.ietf.org/html/rfc7231#section-7.1.3
      schema:
        format: int32
        maximum: 186400
        minimum: 0
        type: integer
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        format: int32
        maximum: 1000000
        minimum: 0
        type: integer
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        format: int32
        maximum: 1000000
        minimum: 0
        type: integer
    X-RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        format: int32
        maximum: 186400
        minimum: 0
        type: integer
  schemas:
    OASDocument:
      type: string
      description: |-
        An OpenAPI Specification Document
        serialized as YAML or JSON.
      format: binary
      maxLength: 1000000
      minLength: 100
      externalDocs:
        url: https://raw.githubusercontent.com/OAI/spec.openapis.org/refs/heads/main/oas/3.0/schema/2024-10-18
        description: The OpenAPI 3.0 JSON Schema
    CalculateSemanticScoreRequest:
      type: object
      additionalProperties: false
      properties:
        file:
          $ref: "#/components/schemas/OASDocument"
      required:
        - file
    Problem:
      additionalProperties: false
      required:
        - title
      properties:
        detail:
          description: |
            A human-readable explanation specific to this occurrence of the
            problem. You MUST NOT expose internal information, personal
            data or implementation details through this field.
          example: Request took too long to complete.
          type: string
          maxLength: 512
          pattern: ^
        instance:
          description: |
            An absolute URI that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced.
          format: uri
          type: string
          maxLength: 1024
        status:
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          example: 503
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: |
            A short summary of the problem type. Written in English and readable
            for engineers (usually not suited for non-technical stakeholders and
            not localized);
          example: Service Unavailable
          type: string
          maxLength: 512
        type:
          default: about:blank
          description: |
            An absolute URI that identifies the problem type.  When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          example: https://tools.ietf.org/html/rfc7231#section-6.6.4
          format: uri
          type: string
          maxLength: 1024
      type: object
    SemanticScoreResponse:
      type: object
      required:
        - score
        - timestamp
        - models
      properties:
        score:
          type: number
          description: Overall semantic score for the schema (0-1)
          example: 0.75
          minimum: 0
          maximum: 1
          format: float
        timestamp:
          type: integer
          description: Unix timestamp of when the score was calculated
          example: 1707234567890
          format: int64
        sparql_endpoint:
          type: string
          description: |-
            SPARQL endpoint URL used for validation.
            This field is used only for debugging purposes and may not be available.
          example: https://example.com/sparql
          format: uri
          deprecated: true
        models:
          type: array
          maxItems: 1000
          minItems: 0
          description: Per-model details with their semantic score
          items:
            $ref: "#/components/schemas/ModelSummary"
    ModelSummary:
      type: object
      required:
        - name
        - score
        - has_annotations
        - raw_properties_count
        - valid_properties_count
        - invalid_properties_count
        - properties
      properties:
        name:
          type: string
          description: >-
            Name of the data model relative to the OpenAPI specification document #/components/schemas path.
          example: Country
        score:
          type: number
          description: Semantic score for this model (0-1)
          example: 0.75
          minimum: 0
          maximum: 1
          format: float
        has_annotations:
          type: boolean
          description: Whether the model is annotated with x-jsonld-context or not.
          example: true
        raw_properties_count:
          type: integer
          description: Number of properties in the model
          example: 10
          minimum: 0
          format: int32
        valid_properties_count:
          type: integer
          description: Number of valid properties in the model
          example: 8
          minimum: 0
          format: int32
        invalid_properties_count:
          type: integer
          description: Number of invalid properties in the model
          example: 2
          minimum: 0
          format: int32
        properties:
          type: array
          maxItems: 1000
          description: List of properties with their validation status
          items:
            $ref: "#/components/schemas/PropertySummary"
    PropertySummary:
      type: object
      required:
        - name
        - valid
      properties:
        name:
          type: string
          description: Name of the property
          example: country
        uri:
          type: string
          description: URI of the property
          example: https://w3id.org/italia/onto/CLV/country
          format: uri
          nullable: true
        valid:
          type: boolean
          description: Whether the property is valid
          example: true